--- index_old.html
@@ -1,14 +1,0 @@
-<!-- index.html -->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Your HTML Document</title>
-    <script type="module" src="./js/file2.js"></script>
-</head>
-<body>
-    <!-- Display the greeting in this element -->
-    <div id="greeting"></div>
-</body>
-</html>
